Updated DPS Scene - Level plumbing complete - Results of game plumbing complete - Countdown timer at beginning complete - Results text at end complete - Speed of character now a function of difficulty

+++ Assets/Scripts/Character.cs
@@ -161,7 +161,7 @@ public class Character : MonoBehaviour
 		return (Character.CharacterEyes)eyes.GetValue(eyeValue);
 	}
 
-    public void MoveAcrossScreen(int force)
+    public void MoveAcrossScreen(float force)
     {
         var multiplier = Random.Range(0.75f, 1.5f);
         rigidBody.AddForce(new Vector2(force * multiplier, 0));
